+++ firmware/application/freertos/task_pose_ctrl.cpp
@@ -18,7 +18,7 @@ static TaskHandle_t task_handle;
 
 static constexpr vPose velocity_smoothen(const vPose& vel_target,
                                          const vPose& vel_cur) {
-  constexpr auto MAX_DIFF_LINEAR = MAX_VELOCITY_WHEEL_LINEAR * 0.03;
+  constexpr auto MAX_DIFF_LINEAR = MAX_VELOCITY_WHEEL_LINEAR * 0.03f;
   constexpr auto MAX_DIFF_ANGULAR = MAX_DIFF_LINEAR / L_W_HALF;
 
   auto vel_diff = vel_target - vel_cur;
@@ -30,15 +30,15 @@ static constexpr vPose velocity_smoothen(const vPose& vel_target,
   return vel_cur + vel_diff;
 }
 
-static Pose pose_ctrl(const Pose& pose_sp, const Pose& pose_cur, double dt) {
+static Pose pose_ctrl(const Pose& pose_sp, const Pose& pose_cur, float dt) {
   return {};
-  static constexpr double MAX_INTEGRAL_LINEAR = 200;
-  static constexpr double MAX_INTEGRAL_ANGULAR = M_PI;
-  static constexpr double MAX_POSE_DEVIATION_LINEAR = 300;
-  static constexpr double MAX_POSE_DEVIATION_ANGULAR = pi;
-  static constexpr double K_P = 0.07;
-  static constexpr double K_I = 0.001;
-  static constexpr double K_D = 0.00001;
+  static constexpr float MAX_INTEGRAL_LINEAR = 200;
+  static constexpr float MAX_INTEGRAL_ANGULAR = M_PI;
+  static constexpr float MAX_POSE_DEVIATION_LINEAR = 300;
+  static constexpr float MAX_POSE_DEVIATION_ANGULAR = pi;
+  static constexpr float K_P = 0.07;
+  static constexpr float K_I = 0.001;
+  static constexpr float K_D = 0.00001;
   static auto integral = Pose{}, prev_err = Pose{};
 
   const auto err = pose_sp - pose_cur;
@@ -48,14 +48,14 @@ static Pose pose_ctrl(const Pose& pose_sp, const Pose& pose_cur, double dt) {
       abs(err.theta) > MAX_POSE_DEVIATION_ANGULAR) [[unlikely]]
     ULOG_WARNING("[pose_ctrl] sanity check: pose deviation too large");
   // ULOG_DEBUG("[pose_ctrl] delta: [x: %.2f, y: %.2f, theta: %.2f]", err.x, err.y,
-  //            static_cast<double>(err.theta));
+  //            static_cast<float>(err.theta));
 
   integral += err * dt;
   integral.x =
       std::clamp(integral.x, -MAX_INTEGRAL_LINEAR, MAX_INTEGRAL_LINEAR);
   integral.y =
       std::clamp(integral.y, -MAX_INTEGRAL_LINEAR, MAX_INTEGRAL_LINEAR);
-  integral.theta = std::clamp(static_cast<double>(integral.theta),
+  integral.theta = std::clamp(static_cast<float>(integral.theta),
                               -MAX_INTEGRAL_ANGULAR, MAX_INTEGRAL_ANGULAR);
 
   const auto derivative = err - std::exchange(prev_err, err);
@@ -85,7 +85,7 @@ static void task_impl(void*) {
     pose_sp += Pose(vRF2vWF(rvel_sp, pose_sp.theta) * dt);
 
     // TODO epsilon
-    constexpr double K_eps = -0.2;
+    constexpr float K_eps = -0.2;
     const auto eps = 0;
 
     const auto rvel_cv =
