+++ firmware/application/pose_types.hpp
@@ -10,9 +10,9 @@ using std::numbers::pi;
 /* Heading of a mobile robot in the range of -pi ... +pi */
 class Heading {
  private:
-  double theta{};
+  float theta{};
 
-  static constexpr double maxPI(double theta) {
+  static constexpr float maxPI(float theta) {
     while (theta > pi) theta -= 2 * pi;
     while (theta < -pi) theta += 2 * pi;
     return theta;
@@ -21,45 +21,45 @@ class Heading {
  public:
   constexpr Heading() = default;
 
-  constexpr Heading(double theta) { this->theta = maxPI(theta); }
+  constexpr Heading(float theta) { this->theta = maxPI(theta); }
 
-  constexpr Heading& operator=(const double val) {
+  constexpr Heading& operator=(const float val) {
     this->theta = maxPI(val);
     return *this;
   }
 
-  constexpr Heading& operator+=(const double& rhs) {
+  constexpr Heading& operator+=(const float& rhs) {
     theta = maxPI(this->theta + rhs);
     return *this;
   }
 
-  constexpr Heading& operator-=(const double divisor) {
+  constexpr Heading& operator-=(const float divisor) {
     this->theta = maxPI(this->theta - divisor);
     return *this;
   }
 
-  constexpr Heading operator+(double rhs) const { return this->theta + rhs; }
+  constexpr Heading operator+(float rhs) const { return this->theta + rhs; }
 
-  constexpr Heading operator-(double rhs) const { return this->theta - rhs; }
+  constexpr Heading operator-(float rhs) const { return this->theta - rhs; }
 
-  constexpr Heading operator*(double factor) const {
+  constexpr Heading operator*(float factor) const {
     return this->theta * factor;
   }
 
-  constexpr operator double() const { return this->theta; }
+  constexpr operator float() const { return this->theta; }
 };
 
 class Pose;
 
 class vPose {
  public:
-  double vx{};
-  double vy{};
-  double omega{};
+  float vx{};
+  float vy{};
+  float omega{};
 
   constexpr vPose() = default;
 
-  constexpr vPose(double vx, double vy, double omega)
+  constexpr vPose(float vx, float vy, float omega)
       : vx{vx}, vy{vy}, omega{omega} {}
 
   explicit vPose(const Pose& pose);
@@ -75,7 +75,7 @@ class vPose {
     return *this;
   }
 
-  friend constexpr vPose operator*(const vPose& lhs, const double factor) {
+  friend constexpr vPose operator*(const vPose& lhs, const float factor) {
     return {lhs.vx * factor, lhs.vy * factor, lhs.omega * factor};
   }
 
@@ -88,27 +88,27 @@ class vPose {
   }
 
   /* transformation of velocities from robot frame into world frame  */
-  friend constexpr vPose vRF2vWF(vPose vRF, double theta) {
-    return {vRF.vx * cos(theta) - vRF.vy * sin(theta),
-            vRF.vx * sin(theta) + vRF.vy * cos(theta), vRF.omega};
+  friend constexpr vPose vRF2vWF(vPose vRF, float theta) {
+    return {vRF.vx * std::cos(theta) - vRF.vy * std::sin(theta),
+            vRF.vx * std::sin(theta) + vRF.vy * std::cos(theta), vRF.omega};
   }
 
   /* transformation of velocities from world frame into robot frame  */
-  friend constexpr vPose vWF2vRF(vPose vWF, double theta) {
-    return {vWF.vx * cos(theta) + vWF.vy * sin(theta),
-            -vWF.vx * sin(theta) + vWF.vy * cos(theta), vWF.omega};
+  friend constexpr vPose vWF2vRF(vPose vWF, float theta) {
+    return {vWF.vx * std::cos(theta) + vWF.vy * std::sin(theta),
+            -vWF.vx * std::sin(theta) + vWF.vy * std::cos(theta), vWF.omega};
   }
 };
 
 class Pose {
  public:
-  double x{};
-  double y{};
+  float x{};
+  float y{};
   Heading theta{};
 
   constexpr Pose() = default;
 
-  constexpr Pose(double x, double y, Heading theta)
+  constexpr Pose(float x, float y, Heading theta)
       : x{x}, y{y}, theta{theta} {}
 
   explicit Pose(const vPose& v_pose)
@@ -133,26 +133,26 @@ class Pose {
     return {lhs.x - rhs.x, lhs.y - rhs.y, lhs.theta - rhs.theta};
   }
 
-  friend constexpr Pose operator*(const Pose& lhs, const double factor) {
+  friend constexpr Pose operator*(const Pose& lhs, const float factor) {
     return {lhs.x * factor, lhs.y * factor, lhs.theta * factor};
   }
 
-  friend constexpr Pose operator/(const Pose& lhs, const double divisor) {
+  friend constexpr Pose operator/(const Pose& lhs, const float divisor) {
     if (divisor == 0.0)
       for (;;);
     return {lhs.x / divisor, lhs.y / divisor, lhs.theta / divisor};
   }
 
   /* transformation of small movements from robot frame into world frame  */
-  friend constexpr Pose pRF2pWF(const Pose& dpose, double th) {
-    return {dpose.x * cos(th) - dpose.y * sin(th),
-            dpose.x * sin(th) + dpose.y * cos(th), dpose.theta};
+  friend constexpr Pose pRF2pWF(const Pose& dpose, float th) {
+    return {dpose.x * std::cos(th) - dpose.y * std::sin(th),
+            dpose.x * std::sin(th) + dpose.y * std::cos(th), dpose.theta};
   }
 
   /* transformation of small movements from world frame into robot frame  */
-  friend constexpr Pose pWF2pRF(const Pose& dpose, double th) {
-    return {dpose.x * cos(th) + dpose.y * sin(th),
-            -dpose.x * sin(th) + dpose.y * cos(th), dpose.theta};
+  friend constexpr Pose pWF2pRF(const Pose& dpose, float th) {
+    return {dpose.x * std::cos(th) + dpose.y * std::sin(th),
+            -dpose.x * std::sin(th) + dpose.y * std::cos(th), dpose.theta};
   }
 };
